8 submission

+++ src/main/java/lang/print/gaps/task4/SettingTypes.java
@@ -8,6 +8,6 @@ public class SettingTypes {
        double forth = 12.90434323324342;
        float   fifth = 12.3F;
        short sixth = 1;
-       int  seventh = 129;
+       short  seventh = 129;
     }
 }
